ecom/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.2.5
attrs==19.3.0
autopep8==1.4.4
backoff==1.6.0
beautifulsoup4==4.8.2
black==19.3b0
boto3==1.11.9
botocore==1.14.9
bs4==0.0.1
certifi==2019.6.16
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.8.0
django-debug-toolbar==2.2
django-dotenv==1.4.2
django-ipware==2.1.0
django-klingon==0.0.8
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-robots==3.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
google-images-download==2.8.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
monotonic==1.5
mypy==0.761
mypy-extensions==0.4.3
numpy==1.18.2
oauthlib==3.0.1
pandas==1.0.3
pep8==1.7.1
Pillow==6.2.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.3.2
selenium==3.141.0
sentry-sdk==0.14.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
stripe==2.38.0
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.6
wrapt==1.11.2

